refactor: mv Download functions to Downloads mod
